VTU FORUM CSHARP FILE
It is the back end code of this forum . which is been written in csharp language.Here the main thing is you have to create your own google API to redirect this page to google login. For the correct authentication you have to register you local machine to this api. After writting this code to the back end you will be Authenticated successfully. And can starts the discussion in the vtu forum